Use absolute imports in main bleachbit code
