+++ bleachbit.py
@@ -18,13 +18,11 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-
-from __future__ import print_function
-
 """
 Launcher
 """
 
+from __future__ import absolute_import, print_function
 
 import os
 import sys
