+++ bleachbit/CLI.py
@@ -18,27 +18,23 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-
-from __future__ import print_function
-
 """
 Command line interface
 """
 
+from __future__ import absolute_import, print_function
+
+from bleachbit.Cleaner import backends, create_simple_cleaner, register_cleaners
+from bleachbit import _, APP_VERSION, encoding
+from bleachbit import Diagnostic, Options, Worker
 
 import logging
 import optparse
 import os
 import sys
 
-from Cleaner import backends, create_simple_cleaner, register_cleaners
-from Common import _, APP_VERSION, encoding
-import Diagnostic
-import Options
-import Worker
-import Common
-
 logger = logging.getLogger(__name__)
+
 
 class CliCallback:
     """Command line's callback passed to Worker"""
@@ -183,7 +179,7 @@ This is free software: you are free to change and redistribute it.
 There is NO WARRANTY, to the extent permitted by law.""" % APP_VERSION)
         sys.exit(0)
     if 'nt' == os.name and options.update_winapp2:
-        import Update
+        from bleachbit import Update
         logger.info("Checking online for updates to winapp2.ini")
         Update.check_updates(False, True,
                              lambda x: sys.stdout.write("%s\n" % x),
@@ -210,7 +206,7 @@ There is NO WARRANTY, to the extent permitted by law.""" % APP_VERSION)
         sys.exit(0)
     if options.gui:
         import gtk
-        import GUI
+        from bleachbit import GUI
         shred_paths = args if options.shred else None
         GUI.GUI(uac=not options.no_uac,
                 shred_paths=shred_paths, exit=options.exit)
